Refactor mmtkObjectBarrier
