+++ openjdk/cpu/x86/mmtkObjectBarrierSetAssembler_x86.cpp
@@ -1,0 +1,67 @@
+#include "precompiled.hpp"
+#include "mmtkObjectBarrier.hpp"
+#include "runtime/interfaceSupport.inline.hpp"
+
+#define __ masm->
+
+void MMTkObjectBarrierSetAssembler::object_reference_write_post(MacroAssembler* masm, DecoratorSet decorators, Address dst, Register val, Register tmp1, Register tmp2) const {
+  if (can_remove_barrier(decorators, val, /* skip_const_null */ true)) return;
+#if MMTK_ENABLE_BARRIER_FASTPATH
+  Label done;
+
+  Register tmp3 = rscratch1;
+  Register tmp4 = rscratch2;
+  Register obj = dst.base();
+  assert_different_registers(obj, tmp2, tmp3);
+  assert_different_registers(tmp4, rcx);
+
+  // tmp2 = load-byte (SIDE_METADATA_BASE_ADDRESS + (obj >> 6));
+  __ movptr(tmp3, obj);
+  __ shrptr(tmp3, 6);
+  __ movptr(tmp2, SIDE_METADATA_BASE_ADDRESS);
+  __ movb(tmp2, Address(tmp2, tmp3));
+  // tmp3 = (obj >> 3) & 7
+  __ movptr(tmp3, obj);
+  __ shrptr(tmp3, 3);
+  __ andptr(tmp3, 7);
+  // tmp2 = tmp2 >> tmp3
+  __ movptr(tmp4, rcx);
+  __ movl(rcx, tmp3);
+  __ shrptr(tmp2);
+  __ movptr(rcx, tmp4);
+  // if ((tmp2 & 1) == 1) goto slowpath;
+  __ andptr(tmp2, 1);
+  __ cmpptr(tmp2, 1);
+  __ jcc(Assembler::notEqual, done);
+
+  __ movptr(c_rarg0, obj);
+  __ lea(c_rarg1, dst);
+  if (val == noreg) {
+    __ movptr(c_rarg2, NULL_WORD);
+  } else {
+    __ movptr(c_rarg2, val);
+  }
+  __ call_VM_leaf_base(FN_ADDR(MMTkBarrierSetRuntime::object_reference_write_slow_call), 3);
+
+  __ bind(done);
+#else
+  __ movptr(c_rarg0, obj);
+  __ lea(c_rarg1, dst);
+  __ movptr(c_rarg2, val == noreg ?  (int32_t) NULL_WORD : val);
+  __ call_VM_leaf_base(FN_ADDR(MMTkBarrierSetRuntime::object_reference_write_post_call), 3);
+#endif
+}
+
+void MMTkObjectBarrierSetAssembler::arraycopy_epilogue(MacroAssembler* masm, DecoratorSet decorators, BasicType type, Register src, Register dst, Register count) {
+  const bool dest_uninitialized = (decorators & IS_DEST_UNINITIALIZED) != 0;
+  if ((type == T_OBJECT || type == T_ARRAY) && !dest_uninitialized) {
+    __ pusha();
+    __ movptr(c_rarg0, src);
+    __ movptr(c_rarg1, dst);
+    __ movptr(c_rarg2, count);
+    __ call_VM_leaf_base(FN_ADDR(MMTkBarrierSetRuntime::object_reference_array_copy_post_call), 3);
+    __ popa();
+  }
+}
+
+#undef __
