+++ openjdk/cpu/zero/mmtkObjectBarrierSetAssembler_zero.hpp
@@ -1,0 +1,5 @@
+#ifndef MMTK_OPENJDK_MMTK_OBJECT_BARRIER_SET_ASSEMBLER_ZERO_HPP
+#define MMTK_OPENJDK_MMTK_OBJECT_BARRIER_SET_ASSEMBLER_ZERO_HPP
+
+class MMTkObjectBarrierSetAssembler;
+#endif // MMTK_OPENJDK_MMTK_OBJECT_BARRIER_SET_ASSEMBLER_ZERO_HPP
